Add OmniFocus-related functionality.
There's now a generic "omnifocus" query that can be added to any
curator. It's not specific to any vault. It is still specific to my
machine though, so it definitely still needs some work.

The functionality pulls projects from OmniFocus and compares them to
pages that represent notes. It can then produce a report of mismatches:

- OmniFocus projects that do not have a matching project page.
- Project pages that do not have a matching OmniFocus project.