Protect against multiple invocation of the application.
The Application class now first creates a PID file. If not possible,
it exits immediately. The PID file is removed on shutdown.

The PID file is written in the temporary directory, and not in /var/run.

Writing in /var/run works on macOS, but not on most Linuxes, and
definitely not on Windows.