Protect against unbalanced/nested query blocks.
Situation:

- a document already contains a query
- the user wants  to insert a new query above it
- the user types the query opening tag
- the user or the app saves the document

In this situation the parser would see a single query, starting at the
new query opening tag, ending at the
query closing tag at the bottom of the second query. Result: everything
in between is overwritten. Not great.

The parser is now more robust: if, while reading a query, it detects
another query opening tag, it will create a textblock instead from the
query it encountered so far, in effect ignoring the query.

Result: fewer surprises!

This bit me in the *ss a few times, because Obsidian likes to save its
files every microsecond or so...